(PC-37176)[PRO] refactor: rename utils as serializers
